requirements.txt
asgiref==3.11.1
Deprecated==1.3.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
gunicorn==25.3.0
img2pdf==0.6.3
lxml==6.0.2
packaging==26.0
pikepdf==10.5.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyMuPDF==1.27.2.2
sqlparse==0.5.5
tzdata==2025.3
wrapt==2.1.2
